Added header image and broke sections into components
