+++ src/Components/About.js
@@ -1,0 +1,11 @@
+// About Section
+const About = () => {
+    return(
+        <div className="about">
+            <h2>How to get started</h2>
+            <p>Plan your week ahead of time with the <strong>Fitness Planner</strong> app! Start by putting in the day of the week(Mon-Sun), then any form of excercise ranging from yoga to the gym to outdoor sports, and follow it up with the length of time you aim to workout for. Afterwards, when the week has ended you can clear the entries and log in your plans for the new week ahead of you. Achieve your goals or strive for a healthy lifetyle!</p>
+        </div>
+    );
+}
+
+export default About;
