+++ src/Components/Footer.js
@@ -1,0 +1,10 @@
+// Footer Section
+const Footer = () => {
+    return(
+        <footer>
+            <p>🏅 Created at Juno College of Technology🏅</p>
+        </footer>
+    );
+}
+
+export default Footer;
