+++ src/Components/Header.js
@@ -1,0 +1,14 @@
+// Header Image
+import header from "../assets/header.jpg";
+// Header Section
+const Header = () => {
+    return(
+        <header>
+            <h1>Fitness Planner</h1>
+            <img src={header} alt="Medicine ball training" />
+        </header>
+    );
+    
+}
+
+export default Header;
